index: Add blog posts titles to show what is possible
This looks ugly to make it clear that we don't actually want to leave it as
it is. Also, we probably want to make separate pages out of each post, and
not render a list, like this. This tutorial explains how to go about doing
this: https://www.gatsbyjs.org/tutorial/part-seven/