Push current HEAD to master from TravisCI
TravisCI does a checkout to the current git sha of the top commit on master.
This means the ref for master is no longer updated on TravisCI, even when a
new commit is made. So, we change the push_to_git.sh script to push the
current HEAD to master, without bothering about the master ref on Travis.